refactor(perf): split /about + /music-lab into server shell + motion island
27th + 28th in the pattern series.
